+++ indra/llimage/llimagefilter.cpp
@@ -507,7 +507,7 @@ void LLImageFilter::filterScreen(EScreenMode mode, const F32 wave_length, const 
     std::array<U8, 256> gamma;
     for (S32 i = 0; i < 256; i++)
     {
-        F32 gamma_i = llclampf((float)(powf((float)(i)/255.0f,1.0f/4.0f)));
+        F32 gamma_i = llclampf(powf((float)(i)/255.0f,1.0f/4.0f));
         gamma[i] = (U8)(255.0 * gamma_i);
     }
 
@@ -756,7 +756,7 @@ void LLImageFilter::filterGamma(F32 gamma, const LLColor3& alpha)
 
     for (S32 i = 0; i < 256; i++)
     {
-        F32 gamma_i = llclampf((float)(powf((float)(i)/255.0f,1.0f/gamma)));
+        F32 gamma_i = llclampf(powf((float)(i)/255.0f,1.0f/gamma));
         // Blend in with alpha values
         gamma_red_lut[i]   = (U8)((1.0f - alpha.mV[0]) * (float)(i) + alpha.mV[0] * 255.0f * gamma_i);
         gamma_green_lut[i] = (U8)((1.0f - alpha.mV[1]) * (float)(i) + alpha.mV[1] * 255.0f * gamma_i);
